refactor(app): extract detectSubmitOwnership into submit-ownership.ts
Pure extraction, no behavior change. Move the SubmitOwnership discriminated
union and the detectSubmitOwnership pure function out of submit.ts into a
focused submit-ownership.ts that depends only on type-level imports
(PromptRouteScope / PortableDraftOwner / PinnedDraftOwner).

submit.ts now imports detectSubmitOwnership for internal use instead of
defining and exporting it. Repoint submit-ownership.test.ts and
draft-isolation.integration.test.ts to import detectSubmitOwnership statically
from the new module; both previously pulled it via dynamic import("./submit")
guarded by a @solidjs/router mock that existed only because submit.ts loads
the router at module scope. With the function isolated, that router mock is
dead and removed; the encode mocks are kept since owners/buildRequestParts
still consume them.

Verified: bun run typecheck (8/8), bun test prompt-input (382 pass, same 2
pre-existing isPromptEqual failures as dev), eslint clean.
